update cli to not step on itself
