+++ lockhart/cli/__init__.py
@@ -3,13 +3,11 @@
 # SPDX-License-Identifier: MIT
 
 
-from rich.console import Console
 import typer
 
-from lockhart import prompts
-from lockhart.config import config as configuration
-from lockhart.console import console
-from lockhart.history import load_history
+from lockhart.cli.config import config_app
+from lockhart.cli.history import history_app
+from lockhart.cli.prompt import prompt_app
 
 app = typer.Typer(
     name="lockhart",
@@ -46,62 +44,6 @@ def main(
     return
 
 
-config_app = typer.Typer()
-prompt_app = typer.Typer()
-history_app = typer.Typer()
 app.add_typer(config_app)
 app.add_typer(prompt_app)
 app.add_typer(history_app)
-
-
-@config_app.callback()
-def config():
-    "configuration cli"
-
-
-@config_app.command()
-def show():
-    console.print(configuration)
-
-
-@prompt_app.callback()
-def prompt():
-    "prompt cli"
-
-
-@prompt_app.command()
-def list():
-    for prompt in configuration.get("prompts", {}).keys():
-        console.print(prompt)
-
-
-@prompt_app.command()
-def run(
-    prompt: str = typer.Argument(..., help="the configured prompt to run"),
-    dry_run: bool = typer.Option(False, help="run without sending the prompt"),
-    verbose: bool = typer.Option(False, help="show the log messages"),
-    edit: bool = typer.Option(
-        False, help="open the resulting prompt in your editor before running"
-    ),
-):
-    if not verbose:
-        console.quiet = True
-    console.log(f"running configured prompt: {prompt}")
-    result = prompts.run_configured_prompt(prompt, dry_run, edit)
-    if result:
-        try:
-            Console().print(result["choices"][0]["text"])
-        except KeyError:
-            Console().print(result)
-
-
-@history_app.callback()
-def history():
-    "history cli"
-
-
-@history_app.command()
-def list(
-    verbose: bool = typer.Option(False, help="show the log messages"),
-):
-    Console().print(load_history())
